+++ ui/src/components/tree/Tree.vue
@@ -81,7 +81,7 @@ export default {
   beforeDestroy () {
     if (!this.whakapapaView) return
     if (this.whakapapaView.name === 'Loading') return
-    if (this.whakapapaView.canEdit) return
+    if (!this.whakapapaView.canEdit) return
 
     if (!this.whakapapaView.id) {
       console.warn('Trying to save the record count without a whakapapa id', this.whakapapaView)